Version 1.2
Added support for multiple arrays as well as a JMenu to indicate whether
the inputfield has been saved or not.

It seems to be that the counting of waste might be faulty but there is
no indication of it in the code, only that it doesn't double evenly with
arrays or decrease. Otherwise the math all seems to be in perfect
working order.